requirements.txt
Flask==3.0.0
cryptography==41.0.7
supabase==2.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
redis==5.0.1
httpx==0.26.0
